MODUSERSKC-154: Users expiration field not working for eureka (#232)
* refactor(kafka): replace ResourceEvent with SystemUserEvent and introduce SystemUser model

* feat(kafka): add user event handling and configuration for user topic

* feat(user): introduce UserDomainEvent model and related Kafka topic configuration

* feat(user): enhance user event handling and logging configuration

* feat(kafka): add concurrency configuration for user topic listener

* feat(user): implement user event handling in KafkaMessageListener and add UserEventDeserializer tests

* feat(user): add user activation and update event handling in integration tests

* feat(kafka): enforce non-null event type in Kafka message listeners and improve logging

* docs: update NEWS.md for version v4.1.0 with users expiration field fix

* feat(docs): update README.md with user topic configuration details

* feat(kafka): enhance user event handling with detailed documentation and retry configuration

* feat(kafka): improve userId retrieval with safe extraction method and enhance test verifications

* feat(kafka): refactor userId retrieval to use User DTO for safer extraction

+++ src/main/java/org/folio/uk/configuration/RetryConfiguration.java
@@ -2,7 +2,6 @@ package org.folio.uk.configuration;
 
 import java.util.List;
 import lombok.extern.log4j.Log4j2;
-import org.folio.uk.integration.kafka.configuration.SystemUserRoleRetryConfiguration;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -37,7 +36,7 @@ public class RetryConfiguration {
   @Bean(name = "systemUserRoleRetryTemplate")
   public RetryTemplate capabilityRetryTemplate(
     @Qualifier("methodLoggingRetryListener") RetryListener retryListener, SystemUserRoleRetryConfiguration config) {
-    return new RetryTemplateBuilder().maxAttempts(config.getRetryAttempts())
+    return new RetryTemplateBuilder().maxAttempts((int) config.getRetryAttempts())
       .fixedBackoff(config.getRetryDelay().toMillis())
       .withListener(retryListener)
       .retryOn(List.of(RestClientResponseException.class))
